Starting project.
I did a little work without version control and it was approaching the
point where I thought "If I lost this, I would be really upset".
So that brought me to this point. To much of this project is
dysfunctional in this project so I don't dare make it public yet.